setting-iframe: Adjust sectionTitles to sentence case
Title Case can be hard to read and, for example, some
languages (German where capitalization goes for all nouns) even harder
to achieve without problems. Furthermore, this iframe will be seen in
the integrator side and is probably a safer option to make sure
everything is sent out in: Sentence case.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: I5183fd5adaf877b557b4636c7b5bfd1ee8dd81c0