fix: show correct page number
In PDFs, we have some code to render the page numbers based on a fixed
page size when tiles haven't yet loaded

Rather than page 1 being the top, the top page is internally considered
to be page 0. That means that when subtracting 1, we would shift all the
rendered placeholder pages up by one page.

This effect wouldn't show when you were at the top of the document,
because then we would be on 'page 0' and would go into the other branch,
not subtracting 1 and not rendering the pages incorrectly offset. Scroll
down a little, however, and you would find that as soon as the top of
page 2 reached the top of your screen the numbers would jump - page "2"
becoming page "3". Finally, you wouldn't get a page rendered at all
where the last page should be

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I18508e54d66d8f500431cbf78ff390069c56c0bb