feat: send data through the mobile socket
We've been working on adding a MobileSocket: a different method of
sending data to the mobile webviews due to major performance problems in
the FakeWebSocket.

We've constructed all of this in previous commits, and now it's time to
glue it all together!

For iOS, the glue involves an ugly __bridge cast -- this is to avoid
some errors I had when importing the URL protocol handler so as to
declare the types properly. It is still sound as the schemeHandler
property is only ever set to our CoolURLSchemeHandler

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Id3874e692d003e534648d06eeacd19fa9c5f2ce5