feat: create MobileSocket class
This is the final bit of work that needs to be done for the mobile
socket feature before I can glue it all together. There should be
nothing that particularly surprises you here, perhaps excepting the
mentioned // HACK: comment

To explain that in depth, we're going to have to go into how JavaScript
inheritance works

Each object in JavaScript has a prototype (__proto__). If a method isn't
found on the object, JavaScript will look up the prototype chain to find
it. Class methods are put *on the prototype of the object, not on it
directly*.

If you have a class called MobileSocket with a send method, calling send
will check these properties in order:

- MobileSocket.send (not defined)
- MobileSocket.__proto__.send (defined in MobileSocket)

The problem comes when we extend the ProxySocket. As it's the old style,
the "constructor" sets these properties directly. Therefore, calling
send will now incorrectly select the ProxySocket send method:

- MobileSocket.send (defined by ProxySocket constructor)
- MobileSocket.__proto__.send (defined in MobileSocket)

To actually override this, we need to unset the ProxySocket version in
our constructor

- MobileSocket.send (deleted in MobileSocket constructor)
- MobileSocket.__proto__.send (defined in MobileSocket)

---

Finally, as mentioned before, we can't use the ProxySocket for sending
*to* the native parts as Android can't read bodies of intercepted
requests. Therefore, we'll stick to our existing send method as used in
the FakeWebSocket code

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I8107a5bcf79d8a4d88cf38df760045489977d688