Calc: do not reset cursor state on sendInitUNOCommands
if no IdleHandler.activate() will happen after that - we will not
get tiles rendered anymore (until we click somewhere).

That causes a race on load, rarely occuring, but reproducible.

Better to keep that internal state in it's parent CalcTileLayer
and not mofidy anywhere else. Reconnect case is handled by reset
in CalcTileLayer and was reason to introduce problematic call
in Map in commit 423aa0fed6f411378ce3b0116bc90271c4217532
    calc: refresh cursor on reconnect

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I59ebdbaf453d25534ba5146d6f1161ef8946f82c