Simplify and track delta decompression transactions
Add a type to encapsulate raw deltas and associate an ID with them
to allow tracking requests and better reasoning about when requests
are required and in what order they need to be processed.
This opens up the possibility of splitting transactions up between
multiple workers and/or working on transactions progressively and
allowing transactions to be cancelled in-flight.
This also provides us with more warnings if we encounter unexpected
situations, such as non-contiguous delta updates.

Adapt patch to LibreOffice Online.

Signed-off-by: Chris Lord <chris.lord@collabora.com>
Change-Id: I346202a056e4dcec8f8085e8dee7d3cc1ba6cf51