queue addforkit messages if root forkit doesn't exist
Which can happen if the remote font config changes and the primordial
forkit exits to restart, and a request for a new subforkit is issued
while the forkit hasn't restarted yet.

That request was silently dropped. Now instead queue the addforkit to be
issued when the main forkit reappears.

Adapt patch to LibreOffice Online.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I0ae2f2d39076114599976966d6af995bf11b47a9