Set CJK/CTL settings here for LOKit
share/registry/cjk*.xcd and share/registry/ctl*.xcd are not installed for COOL.
Maybe it's a bug, but... These files are in the so called "brand" layer, in
collaboraoffice-<lang> deb/rpm that we do not use.
On the other hand, why would a feature like inserting vertical text
depend on an optional language pack?
It is a valid use case when someone uses English UI and edits CJK/CTL text,
or when someone wants to insert vertical text e.g. in English.

Adapt patch to LibreOffice Online.

Signed-off-by: Andras Timar <andras.timar@collabora.com>
Change-Id: I2ed446aeff8a6a0a00c0a489eaa1d26778e1b61e