iconview: Persist icon-view items as they don't change
In the first rendering we only have .ui-iconview-entry > img but when we persist iconview items and try to render again it adds intermediate div. i.e. .ui-iconview-entry > .ui-iconview-icon > img

Due to this context menu on right click on any layout is opening at the bottom part after all layout image. This commit resolve this issue and   keep the original structure after persist.

Signed-off-by: Parth Raiyani <parth.raiyani@collabora.com>
Change-Id: If4b1b60efb4aed268fe6b8b7a0396e7a41819411