fix(gtkapp): don't attempt to use mobilesocket
This is a partial revert of Id3874e692d003e534648d06eeacd19fa9c5f2ce5.
On iOS and Android this communication method works fine, but there is
another platform which is technically considered "mobile" by the code:
the gtk app. Since as we don't have an implementation of the
cool:/cool/mobilesocket endpoint for it, when we ungated that for all
mobile platforms it started blowing up. That's no good!

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I8efee81f2c436585658d6c4affe91abff0844d74