make: czech target now runs the unit-tests only
Running `make check` is increasingly slow
because of the JS checks and the mocha tests.

While it's good to have those, when one isn't
changing the JS code (or indeed anything in browser/)
this is pure waste of precious time.

Since `make czech` existed and didn't do much,
we repurpose it here to run only the unit-tests.

This reduces the change-build-test cycle by
a huge amount.

Adapt patch to LibreOffice Online.

Change-Id: I8c5f0c9082644c333c344461b88ee625b90e3fea
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>