Revive COWASM: Don't garble Unicode content of "jsdialog: ..." messages
54ad0f173b49028cba48decba1a159a8fb133a99 "Another fix for issue #5843" had fixed
this code for iOS, by restricting assignment to e.imgBytes to only certain
message types, but conservatively left the broken code alone for Emscripten.

But when e.data contains UTF-16 code units > 0xFF, e.imgBytes clipped those to
[0x00..0xFF], and _onMessage then constructed its textMsg from e.imgBytes and
passed that broken textMsg to _onJSDialog.  So trying to show dialogs in COWASM
easily caused the JSON.parse in _onJSDialog to throw a "SyntaxError: Bad control
character in string literal in JSON at ..." exception, when e.g. a U+0410
CYRILLIC CAPITAL LETTER A in the textMsg had been garbled to a U+0010 control
character.

Signed-off-by: Stephan Bergmann <stephan.bergmann@collabora.com>
Change-Id: I13145998c77ed0aa4096130d7e706e9fc03b9a9a