Fix jump when we opened sidebar in Impress
When we opened sidebar in Impress - it jumped because of
scrollIntoView call in Icon view for slide layouts.

Let's do not animate it. It only distracts people while
doesn't give any value. Also I noticed unused animation time
which was replaced with CSS - removed to not confuse developers.
Also Sidebar doesn't have any parameters in different cases...
so I simplified constructor to not pass it over and over again.

Adapt patch to LibreOffice Online.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Ie71c8ef773c1851e570e188baeae2da0362339b5