feat(treeview): allow cancelling focus grabs
Selections cause problems with the new treeview editing code: if we
process a selection while we are editing, we grab focus which blurs the
edit field and stops editing.

To avoid selections causing trouble with editing in our treeview we have
considered disabling selections on editable elements. Unfortunately,
this causes problems if you want to just select an element (rather
than editing it). This way we can granularly disable selection for the
treeview while we are editing it.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ib099cb1610d97215176e735502d619c31f7da2e9