build(iOS): automatically generate version number
We make a lot of version bumping commits, but tend to neglect to
update iOS at the same time. This is a shame, since as iOS won't allow
you to build from a version you have already released from so we are
effectively required to bump versions after releasing anyway.

We can, however, generate the version that I bump to anyway by removing
the first dot from the COOLWSD_VERSION.

There is a minor risk that we may not bump versions for sufficiantly
long that the iOS release "catches up" with the version on master
(or more likely distro/collabora/co-25.04). In this case, we're in a
no-worse situation than we are in now in that we will likely have to
make a bump to the nano version on a separate mobile branch.

This was a much bigger problem before we changed the version format to
take the nano version into account, which seems to (according to the
comment) be why this wasn't done before... Also, we would previously
have had to bump the version in the main distro/collabora/co- branch
whereas now we have a separate mobile branch. All to say: this probably
isn't an issue anymore.

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I6a6a69644e47d79017e50f8916e5293570229221