feat(formulabar): wrap in scrollable element
I'm about to add some scroll handles that use some altogether-too-clever
CSS to pull them out of the stacking context they are in and display
them ontop of some other elements.

To do this, I'll need a wrapper for the formulabar to use as a sort
of hold to grab on to. It needs to be relatively-positioned so we can
make our handle layer position absolutely to it, pulling it out of its
stacking context.

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I6a6a69641f2d29ec32e9f3f54b6484fe27ab5200