wsd: poll timeout must not be longer than lock refresh period
If we poll longer than the lock refresh period,
we end up checking/refreshing too late.

This isn't critical, in practice, but for
tests it does make a difference. And it's
always best to take it into account.

Given that the default refresh period is
900 seconds and the default poll timeout
is 64 seconds, this is only meaningful
for tests (or the default overridden).

Change-Id: I8d4cabdddd727b2b1320f53d3572754c9e151026
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>