jsdialog: prevent adding empty space on top of tabpages
Example was: Writer -> Format -> Page Style -> Area with None
selected. The content was moved to vertical center of a tabpage
causing weird whitespace on top.

when we process tab page content it contains usually container with
single child which we interpreted as grid and also were adding additional
container on top

grid with single child element should be a regular container

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I4f4beb8cb1e32686cbb162bc9477b3fe362abf6b